Remove Filter (group) in favor of more composable component
