+++ src/index.js
@@ -24,7 +24,6 @@ export { default as Select } from './select'
 export { default as Slider } from './slider'
 export { default as Table } from './table'
 export { default as Tag } from './tag'
-export { default as TagOptions } from './tag-options'
 export { default as Toggle } from './toggle'
 export { default as Tracking } from './tracking'
 export { default as Tray } from './tray'
